--- nav2_default_recoveries/recovery_plugins.xml
@@ -1,15 +1,0 @@
-<class_libraries>
-  <library path="nav2_default_recoveries">
-    <class name="nav2_recoveries/Spin" type="nav2_recoveries::Spin" base_class_type="nav2_core::Recovery">
-      <description>The spin recovery</description>
-    </class>
-
-    <class name="nav2_recoveries/BackUp" type="nav2_recoveries::BackUp" base_class_type="nav2_core::Recovery">
-      <description>The back up recovery</description>
-    </class>
-
-    <class name="nav2_recoveries/Wait" type="nav2_recoveries::Wait" base_class_type="nav2_core::Recovery">
-      <description>The wait recovery</description>
-    </class>
-  </library>
-</class_libraries>
