--- nav2_recoveries/test/CMakeLists.txt
@@ -1,7 +1,0 @@
-ament_add_gtest(test_recoveries
-  test_recoveries.cpp
-)
-
-ament_target_dependencies(test_recoveries
-  ${dependencies}
-)
