+++ nav2_recovery/test/CMakeLists.txt
@@ -1,0 +1,7 @@
+ament_add_gtest(test_recovery
+  test_recovery.cpp
+)
+
+ament_target_dependencies(test_recovery
+  ${dependencies}
+)
