+++ nav2_system_tests/src/recoveries/spin/spin_recovery_tester.cpp
@@ -75,7 +75,7 @@ void SpinRecoveryTester::activate()
     rclcpp::spin_some(node_);
   }
 
-  // Wait for lifecycle_manager_navigation to activate recoveries_server
+  // Wait for lifecycle_manager_navigation to activate recovery_server
   std::this_thread::sleep_for(10s);
 
   if (!client_ptr_) {
